+++ lib/widgets/home_screen/learn_tab/learn_tab.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:elola/localizations.dart';
 import 'package:elola/widgets/game_screen/game_screen.dart';
 
-class LearnScreen extends StatelessWidget {
-  const LearnScreen({Key key}) : super(key: key);
+class LearnTab extends StatelessWidget {
+  const LearnTab({Key key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
