+++ lib/widgets/my_app.dart
@@ -12,7 +12,7 @@ import 'package:elola/modules/text_to_speech/text_to_speech.dart';
 import 'package:elola/modules/user_settings/user_settings.dart';
 import 'package:elola/utils/hive_utils.dart';
 import 'package:elola/widgets/home_screen/home_screen.dart';
-import 'package:elola/widgets/home_screen/settings_screen/settings_store.dart';
+import 'package:elola/widgets/home_screen/settings_tab/settings_store.dart';
 
 class MyApp extends StatefulWidget {
   @override
